+++ frontend/src/pages/Graphing/components/Graph.tsx
@@ -932,6 +932,7 @@ const Graph = ({
 						spotlight={spotlight}
 						setTimeRange={setTimeRange}
 						loadExemplars={loadExemplars}
+						showGrid
 					>
 						{children}
 					</LineChart>
@@ -949,6 +950,7 @@ const Graph = ({
 						spotlight={spotlight}
 						setTimeRange={setTimeRange}
 						loadExemplars={loadExemplars}
+						showGrid
 					>
 						{children}
 					</BarChart>