+++ frontend/src/pages/Player/styles.css.ts
@@ -67,7 +67,6 @@ export const playerCenterColumn = style({
 	flexGrow: 1,
 	position: 'relative',
 	minWidth: MIN_CENTER_COLUMN_WIDTH,
-	zIndex: 0,
 })
 
 export const draggable = style({