+++ sdk/client/src/utils/sessionStorage/highlightSession.ts
@@ -5,51 +5,51 @@ import { SESSION_PUSH_THRESHOLD } from '../../constants/sessions'
 export type SessionData = {
 	sessionSecureID: string
 	projectID: number
+	payloadID: number
 	sessionStartTime?: number
 	lastPushTime?: number
 	userIdentifier?: string
 	userObject?: Object
 }
 
-const getSessionData = (): SessionData | undefined => {
-	let storedSessionData = JSON.parse(
-		getItem(SESSION_STORAGE_KEYS.SESSION_DATA) || '{}',
-	)
+const getSessionDataKey = (sessionID: string): string => {
+	return `${SESSION_STORAGE_KEYS.SESSION_DATA}_${sessionID}`
+}
+
+export const getSessionSecureID = (): string => {
+	return getItem(SESSION_STORAGE_KEYS.SESSION_ID) ?? ''
+}
+
+export const setSessionSecureID = (secureID: string) => {
+	setItem(SESSION_STORAGE_KEYS.SESSION_ID, secureID)
+}
+
+const getSessionData = (sessionID: string): SessionData | undefined => {
+	const key = getSessionDataKey(sessionID)
+	let storedSessionData = JSON.parse(getItem(key) || '{}')
 	return storedSessionData as SessionData
 }
 
-export const getPreviousSessionData = (): SessionData | undefined => {
-	let storedSessionData = getSessionData()
+export const getPreviousSessionData = (
+	sessionID?: string,
+): SessionData | undefined => {
+	if (!sessionID) {
+		sessionID = getSessionSecureID()
+	}
+	let storedSessionData = getSessionData(sessionID)
 	if (
 		storedSessionData &&
 		storedSessionData.lastPushTime &&
 		Date.now() - storedSessionData.lastPushTime < SESSION_PUSH_THRESHOLD
 	) {
 		return storedSessionData as SessionData
+	} else {
+		removeItem(getSessionDataKey(sessionID))
 	}
 }
 
-export const setSessionData = function (sessionData: SessionData | null) {
-	if (sessionData === null) {
-		// preserve sessionSecureID as that is used by network listeners
-		setItem(
-			SESSION_STORAGE_KEYS.SESSION_DATA,
-			JSON.stringify({
-				sessionSecureID: getSessionData()?.sessionSecureID,
-			}),
-		)
-		return
-	}
-	setItem(SESSION_STORAGE_KEYS.SESSION_DATA, JSON.stringify(sessionData))
-}
-
-export const getSessionSecureID = function () {
-	const data = getSessionData()
-	return data?.sessionSecureID ?? ''
-}
-
-export const setSessionSecureID = function (sessionSecureID: string) {
-	const data = getSessionData() ?? ({ sessionSecureID } as SessionData)
-	data.sessionSecureID = sessionSecureID
-	return setSessionData(data)
+export const setSessionData = function (sessionData?: SessionData) {
+	if (!sessionData?.sessionSecureID) return
+	const secureID = sessionData.sessionSecureID!
+	setItem(getSessionDataKey(secureID), JSON.stringify(sessionData))
 }